refactor: extensionCommon.ts -> extension.ts
Now the extension.ts files will implicitly be "common" files
and extensionNode will contain Node specific code
