+++ packages/core/src/extension.ts
